Rename to ProcessView, and fix SystemViewData file name
